refactor: Extract factory list elements to library folder
