+++ src/syntax/tree/syntaxTree.spec.ts
@@ -17,15 +17,16 @@ import {
 } from './syntaxTree';
 import { getInstance } from '../warehouse/warehouse';
 
-import { ElementProcess, ElementRoutine } from '../elements/elementProgram';
-import { ElementValueBoolean } from '../elements/elementValue';
-import { ElementOperatorMathPlus } from '../elements/elementOperatorMath';
-import { ElementBoxBoolean } from '../elements/elementBox';
 import { TreeNodeData, TreeNodeStatement, TreeNodeBlock } from './node';
 import { ITreeSnapshotInput } from '@/@types/syntaxTree';
 
 import { registerElementSpecificationEntries } from '../specification/specification';
-import elementSpecificationEntries from '../specification/specificationEntries';
+import elementSpecificationEntries from '../../library/specification';
+
+import { ElementProcess, ElementRoutine } from '@/library/elements/elementProgram';
+import { ElementValueBoolean } from '@/library/elements/elementValue';
+import { ElementOperatorMathPlus } from '@/library/elements/elementOperatorMath';
+import { ElementBoxBoolean } from '@/library/elements/elementBox';
 
 // -------------------------------------------------------------------------------------------------
